refactor: Improve code organization and project structure
BREAKING CHANGES:
- Moved server.js to backend-auth/server.js for consistency
- Updated package.json bin reference to new location

Changes:
- Extracted routes into separate files (backend-*/routes.js)
- Created scripts/ directory for build utilities
- Created config/ directory with example configuration files
- Added automated setup script (npm run setup)
- Added ESLint configuration and lint commands
- Updated README with clearer setup instructions
- Added ARCHITECTURE.md and CHANGELOG.md documentation

Benefits:
- Consistent backend structure (all in backend-* directories)
- Better separation of concerns (routes vs server initialization)
- Easier first-time setup with example configs
- Improved code quality with linting
- Comprehensive documentation

Migration: Update any scripts referencing 'node server.js' to 'node backend-auth/server.js'
