+++ .eslintrc.js
@@ -1,0 +1,16 @@
+module.exports = {
+  env: {
+    node: true,
+    es6: true,
+  },
+  extends: ['eslint:recommended'],
+  parserOptions: {
+    ecmaVersion: 2020,
+    sourceType: 'module',
+  },
+  rules: {
+    'no-console': 'off',
+    'no-unused-vars': ['warn', { argsIgnorePattern: '^_' }],
+    'prefer-const': 'warn',
+  },
+}
