+++ config/jobs.config.example.js
@@ -1,0 +1,12 @@
+/**
+ * Jobs Backend Configuration
+ * 
+ * Copy this file to ../customisation/jobs.config.js and update with your values
+ */
+const parameters = {
+  JOB_PATH: "C:/Dashboard",
+  MODEL_PATH: "C:/DashboardModel",
+  SERVER_PORT: "4000",
+}
+
+module.exports = parameters
