+++ config/resend-users.config.example.js
@@ -1,0 +1,9 @@
+/**
+ * Resend Users Configuration
+ * 
+ * Array of user names (in capital letters) who may resend messages.
+ * If not given or empty, all authenticated users can resend messages.
+ * 
+ * Copy this file to ../customisation/resend-users.config.js and update with your values
+ */
+module.exports = []
