+++ package.json
@@ -2,7 +2,7 @@
   "name": "soa-dashboard",
   "version": "1.0.0",
   "license": "MIT",
-  "bin": "server.js",
+  "bin": "backend-auth/server.js",
   "dependencies": {
     "@koa/cors": "3.4.1",
     "activedirectory2": "2.1.0",
@@ -17,28 +17,32 @@
     "ramda": "0.28.0"
   },
   "scripts": {
-    "backup:config": "node backupConfig.js",
-    "start:auth": "node server.js",
+    "setup": "node scripts/setup-config.js",
+    "backup:config": "node scripts/backupConfig.js",
+    "lint": "npx eslint backend-*/**/*.js scripts/**/*.js",
+    "lint:fix": "npx eslint backend-*/**/*.js scripts/**/*.js --fix",
+    "start:auth": "node backend-auth/server.js",
     "start:file": "node backend-jobs/server.js",
     "start:frontend": "cd frontend && npm run start",
     "start": "npm-run-all --parallel start:frontend start:auth ncc:server:file --print-label",
-    "ncc:server:auth": "ncc run server.js",
-    "ncc:server:file": "ncc run backend-jobs/server",
-    "ncc:build": "ncc build server.js -o dist/auth && ncc build backend-jobs/server.js -o dist/jobs",
+    "ncc:server:auth": "ncc run backend-auth/server.js",
+    "ncc:server:file": "ncc run backend-jobs/server.js",
+    "ncc:build": "ncc build backend-auth/server.js -o dist/auth && ncc build backend-jobs/server.js -o dist/jobs",
     "build": "cd frontend && npm run build",
     "prepkg:server:dashboard": "npm run build",
-    "pkg:server:dashboard": "pkg . --target node10-win-x64 --output esb-dashboard.exe --debug > output-dashboard-win-pkg.log",
+    "pkg:server:dashboard": "pkg backend-auth/server.js --target node10-win-x64 --output esb-dashboard.exe --debug > output-dashboard-win-pkg.log",
     "pkg:server:file": "pkg ./backend-jobs/server.js --options max_old_space_size=4096 --target node10-win-x64 --output esb-jobs.exe --debug > output-jobs-windows-pkg.log",
     "pkg:all": "npm run pkg:server:dashboard && npm run pkg:server:file",
     "build:all": "npm run backup:config && npm run pkg:all && npm run ncc:build",
     "postbuild:all": "cpy ./dist/auth/index.js ./frontend/build --rename=auth.js",
     "release-it": "cd frontend && npm run release-it && git push --follow-tags",
-    "build-auth": " npx ncc build server.js -o dist/auth",
+    "build-auth": "npx ncc build backend-auth/server.js -o dist/auth",
     "postbuild-auth": "cpy ./dist/auth/index.js ./frontend/build --rename=auth.js --flat"
   },
   "devDependencies": {
     "@vercel/ncc": "0.34.0",
     "cpy-cli": "4.2.0",
+    "eslint": "^8.57.0",
     "npm-run-all": "4.1.5",
     "pkg": "5.8.0"
   },
